authentication reducer reformatted
