+++ src/reducers/index.tsx
@@ -1,40 +1,7 @@
-import { combineReducers, Reducer } from 'redux';
-import { SET_AUTHENTICATION } from '../actions/authenication';
+import { combineReducers } from 'redux';
+import { authentication } from './authentication';
 
-// export const posts = (state=[], action) => {
-//   switch(action.type){
-//     default:
-//       return state
-//   }
-// } 
-
-const AUTHENTICATION_INITIAL_STATE = {
-    user: null,
-    pending: true,
-}
-
-interface AUTH_STATE {
-    user: string | null; // TODO: confirm that this is a string OR is it a boolean? How are you tracking auth success
-    pending: boolean;
-}
-
-interface AUTH_ACTION_TYPE {
-  type: typeof SET_AUTHENTICATION,
-  payload: AUTH_STATE
-}
-
-type ACTION_TYPE = AUTH_ACTION_TYPE
-
-export const authentication : Reducer<AUTH_STATE, ACTION_TYPE> = (state = AUTHENTICATION_INITIAL_STATE, action: AUTH_ACTION_TYPE) : AUTH_STATE=> {
-  switch(action.type){
-    case SET_AUTHENTICATION: 
-      const { user, pending } = action.payload
-      return { user, pending }
-    default:
-      return state
-  }
-}
 
 export default combineReducers({
-    authentication: authentication
+    authentication: authentication,
 })
